Revert "Move custom service-specific functions to own package."
This reverts commit 2d050db0cf29919336f9b1b1898856601763403b.
